requirements.txt
Django==1.7.3
resolver==0.2.1
selector==0.8.11
Paste==1.7.2
PasteDeploy==1.3.3
PasteScript==1.7.3
feedparser==4.1
Markdown==2.3.1
pyenchant==1.5.1
smartypants==1.6.0.2
uuid==1.30
psycopg2==2.5.1
Pillow==2.7.0
nose==1.3.0
coverage==3.5.3
versiontools==1.8.2
pyflakes==0.5.0
pep8==1.4
flake8==1.7.0
logilab-common==0.59.0
logilab-astng==0.24.1
pylint==1.3.0
fuzzywuzzy==0.1
sure==1.1.7
ipython==0.13.1
ipdb==0.7
extras==0.0.3
python-mimeparse==0.1.4
testtools==0.9.35
testscenarios==0.4
python-subunit==0.0.18
lettuce==0.2.19ccnmtl
selenium==2.29.0
python-ldap==2.4.13
lxml==3.2.3
factory_boy==2.1.1
statsd==2.0.2

django-treebeard==2.0b2
django-statsd-mozilla==0.3.8.7
django-annoying==0.8.0
django-smartif==0.1
django-markwhat==1.0
djangowind==0.13.2
sorl==3.1
template_utils==0.4p2
typogrify==1.0
django-indexer==0.2.1
django-templatetag-sugar==0.1
django-paging==0.2.2
raven==4.0.3
django-pagetree==1.0.1
django-pageblocks==1.0.0
django-pagetree_export==0.3
django-nose==1.2
django-jenkins==0.16.3
django-smoketest==0.3.0
django-waffle==0.9.0
